Safari-611.1.10/LayoutTests/http/tests/cache/disk-cache/disk-cache-validation-no-body-expected.txt

Test permutations of various cache headers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 81 tests

--------Testing loads from disk cache--------
response headers: undefined
response source: Network

response headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
response source: Disk cache

response headers: {"Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=0","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(0)"}
response source: Disk cache

response headers: {"Expires":"now(100)"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(100)"}
response source: Disk cache

response headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)"}
response source: Network

response headers: {"ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=0","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=100","ETag":"match"}
response source: Disk cache

response headers: {"Expires":"now(0)","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=0","Expires":"now(0)","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=100","Expires":"now(0)","ETag":"match"}
response source: Disk cache

response headers: {"Expires":"now(100)","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=100","Expires":"now(100)","ETag":"match"}
response source: Disk cache

response headers: {"Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"match"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)","ETag":"match"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"match"}
response source: Disk cache after validation

response headers: {"Expires":"now(0)","Cache-control":"no-cache","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)","ETag":"match"}
response source: Disk cache after validation

response headers: {"Expires":"now(100)","Cache-control":"no-cache","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)","ETag":"match"}
response source: Disk cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)","ETag":"match"}
response source: Disk cache after validation

response headers: {"ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","ETag":"nomatch"}
response source: Disk cache

response headers: {"Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(0)","ETag":"nomatch"}
response source: Disk cache

response headers: {"Expires":"now(100)","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(100)","ETag":"nomatch"}
response source: Disk cache

response headers: {"Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)","ETag":"nomatch"}
response source: Network

--------Testing loads through memory cache (XHR behavior)--------
response headers: undefined
response source: Network

response headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
response source: Memory cache

response headers: {"Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=0","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(0)"}
response source: Memory cache

response headers: {"Expires":"now(100)"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(100)"}
response source: Memory cache

response headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)"}
response source: Network

response headers: {"ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","ETag":"match"}
response source: Memory cache

response headers: {"Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0","Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","Expires":"now(0)","ETag":"match"}
response source: Memory cache

response headers: {"Expires":"now(100)","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","Expires":"now(100)","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"match"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)","ETag":"match"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Expires":"now(0)","Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Expires":"now(100)","Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)","ETag":"match"}
response source: Memory cache after validation

response headers: {"ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","ETag":"nomatch"}
response source: Memory cache

response headers: {"Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(0)","ETag":"nomatch"}
response source: Memory cache

response headers: {"Expires":"now(100)","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(100)","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)","ETag":"nomatch"}
response source: Network

--------Testing loads through memory cache (subresource behavior)--------
response headers: undefined
response source: Network

response headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
response source: Memory cache

response headers: {"Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=0","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(0)"}
response source: Memory cache

response headers: {"Expires":"now(100)"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(100)"}
response source: Memory cache

response headers: {"Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-cache"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)"}
response source: Network

response headers: {"ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","ETag":"match"}
response source: Memory cache

response headers: {"Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0","Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","Expires":"now(0)","ETag":"match"}
response source: Memory cache

response headers: {"Expires":"now(100)","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100","Expires":"now(100)","ETag":"match"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"match"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)","ETag":"match"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)","ETag":"match"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Expires":"now(0)","Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Expires":"now(100)","Cache-control":"no-cache","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)","ETag":"match"}
response source: Memory cache after validation

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)","ETag":"match"}
response source: Memory cache after validation

response headers: {"ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","ETag":"nomatch"}
response source: Memory cache

response headers: {"Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(0)","ETag":"nomatch"}
response source: Memory cache

response headers: {"Expires":"now(100)","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100","Expires":"now(100)","ETag":"nomatch"}
response source: Memory cache

response headers: {"Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-store","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-store","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-store","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(0)","Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(0)","ETag":"nomatch"}
response source: Network

response headers: {"Expires":"now(100)","Cache-control":"no-cache","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=0, no-cache","Expires":"now(100)","ETag":"nomatch"}
response source: Network

response headers: {"Cache-control":"max-age=100, no-cache","Expires":"now(100)","ETag":"nomatch"}
response source: Network

PASS successfullyParsed is true

TEST COMPLETE
